grant rhmi-developer users view access to the Fuse namespace
The rhmi-developers group will be used to grant general permissions to
non-admin users. This change updates the rolebinding created during the
view fuse permissions reconcile to reference the cluster role view which
has all the permissions required to allow view access to the Fuse
namespace for rhmi-developer users.

A new clusterrolebinding is added to the operator rbac file so that the
operator can grant these permissions. This binds the cluster role view to
the operator service account.