🐛 Return new error (wrapping nil returns nil)
The intention of the older version here was to return an error with
some info about the reason. It tried to add that
additional context around an additional error by using the
pkgerr.Wrapf function. Unfortunately err is always nil at this point,
and wrapping nil returns nil; so this would return nil if the key
wasn't found in the data, so it would kind of silently be failing.

Fixing this made the test fail, which was easily fixed by updating the
test data.