set awaiting phase when waiting for alert manager route to exist
currently the alert manager config secret setup attempts to get
the alert manager route as one of it's final steps. however, this
route is not guarenteed to exist when the config secret reconcile
begins.

this commit changes getting the alert manager route to the first
step in the function and does not return a failed phase if the
route is not found.

verification:
- run the operator from a fresh install
- ensure no errors are shown while waiting for the alert manager
  route to exist
- run make test/unit
- ensure the tests pass successfully