feat(rhssouser): Create dedicated-admins and realm-managers groups
Implement functionallity in RHSSO User operator to create a dedicated-admins group
with a realm-managers child group.
The dedicated-admins group has the manage-users and view-realm client roles,
and the realm-managers group has all roles for each realm that is
created except "impersonation".
Implement functionallity to add the dedicated admin generated users from
Openshift to the dedicated-admins and realm-managers groups.

Refactor the logic to reconcile Keycloak groups based on a `keycloakGroupSpec`
struct that contains the desired status of the group.

Update installation controller to enqueue reconcile every 5 minutes in order
to reconcile new realms.