fix(threescale): Remove logic that demotes manually promoted users
Remove logic from `reconcileOpenshiftUsers` function that demotes users that have an admin
role in 3scale but are not admins in Openshift, as those users might have been manually
promoted by an admin.

Extract the logic to sync admin/member users to a separate function `syncOpenshiftAdminMembership`
and write unit tests to cover the change.