🚑 Fix openapi generation in Makefile
The change here is to run openapi-gen to get the
zz_generated.openapi.go files generated as expected. This command
requires a boilerplate file as an argument, so I copied the
default (which doesn't resolve properly...) to
./hack/boilerplate.go.txt. Oddly, "operator-sdk generate k8s" now
magically picks up on that and adds it to the zz_generated.deepcopy.go
files.

I've also added a new pattern here for using Go tool binaries in the
Makefile: if it's on the path, use it from there (faster), otherwise
use it through `go run <pkg>` (slower).

To make sure that the version of the tool used through `go run <pkg>`
is reproducible, I've created a tools.go file -- now when one runs `go
mod tidy`, it will make sure that the package is there, and the
version specified in go.mod will be the version used when running `go
run <pkg>`. That's pretty cool.