🚨 Fix alerts that only group by container
JIRA: https://issues.redhat.com/browse/INTLY-6496

More than one Pod in the namespace could have a shared container name,
leading to incorrect alerts.

For example, if a container was exceeding 90% of its memory limit and
we wanted to alert on that; we wouldn't get alerted if it was summing
the usage and limits of more than one container (unless the sums also
happened to exceed 90% of the total limits for all of those containers
combined).

This change also renames the relevant alerts to make it clearer that
they're per Container rather than per Pod, and adds more context to
the descriptions.