test: Add test for default user email address
Add test to ensure that users from the IDP have an email address
in the KeycloakUser CR that's generated by the operator.

If the user from the IDP has no email address, a dummy one will be
generated as <user name>@rhmi.io. Otherwise, the user email address
will remain:

Create two Openshift users, one with email address and one without it,
and test that the KeycloakUser CRs have the expected email addresses.