feat: Add support for a webhook server running with the operator.
Implement logic to start the webhook server provided by the Manager, and to reconcile
ValidatingWebhookConfiguration CR pointing to the service.
Add a new Service pointing to the operator pod and inject TLS certificates into the pod.
Edit operator role to allow to manage ValidatingWebhookConfiguration CRs.

Implement unit tests for reconcilliation.