Fix the resource usage aggregation and deep linking issue JIRA Link: https://issues.redhat.com/browse/INTLY-7869 JIRA Link: https://issues.redhat.com/browse/INTLY-5962
The cluster grafana , does not have the option of "all" for namespaces. To keep the monitoring
grafana inline with that and solve the deeplinking issue ( "all" namespace has no specific location like a targeted namespace like "redhat-rhmi-fuse").
Also, with all namespace option, the aggregation of the resources show misleading results and usage graphs are not usable. Fixed by removing the
"all" option in the dashboards.