fix(tests): Fix attempt to index first element of possibly empty list
The `get3ScaleAPIcasePod` function in the `TestNetworkPolicyAccessNSToSVC` test retrieves
a list of the APICast pods in the 3scale namespace, and returns the first element on the
list. This has caused some tests to panic, possibly because the test is executed before
any pod has been created.

Modify the `get3ScaleAPIcasePod` function to check the size of the pod list and, if empty,
fail the test.