fix(rhmiconfig): Make maintenance window for RHMIConfig case insensitive
The days difference calculated from a date to the next maintenance window uses a
`map[string]int` with the different week days in lowercase. If the value
from the maintenance window is not in lowercase it will fail silently to
match a value from that `map`, so it'll default to 0 as the zero value
of the map, which is Sunday, resulting in the day difference until the
next Sunday.

Convert the week day to lower case before indexing the map.