feat: Notify solution explorer of available upgrades
Add a parameter to the spec.Template.Parameters field of the solution explorer CR
with information about an incoming upgrades.
Add logic to clear this value when the upgrade is approved
Default the value to `null` in order for the UI to parse the value and check
for nullity.
Modify logic of WebApp reconciliation to reconcile parameters without removing
custom parameters.
Add unit tests.
Include `context.Context` value in subscription controller reconciler and use this
value instead of `context.TODO()` in subsequent calls.