fix rhmiconfig 'wait for maintenance 0 days' test case
currently the 'wait for maintenance 0 days' will fail on sundays
due to the arithmetic in calculating the expectedSchedule value of
the test.

the current process of adding x number of days to the current date
where x is '7-int(now().Weekday())' will result in '7' on sundays
as 'int(now().Weekday())' will resolve to '0' on sundays.

this results setting the expected upgrade date as next sunday,
or 7 days from now, whereas the current logic of rhmiconfig will
wait 0 days.

the change now takes the output of '7-int(now().Weekday())' and
performs a modulo '7' operation upon it, essentially resulting in
0 days being added to the current day on sundays.

verification:
- run 'make test/unit'