fix(rhmiConfig): Remove usage of recalculateSchedule annotation
This annotation is not needed anymore, after the refactor carried for INTLY-8450,
however it's still being set by the mutating webhook, and removed when setting
the upgrade status. However, in the case that the status is updated and no
annotations have been added to the CR, a panic may arise as the Annotations field
is `nil`.
Removing all usage of this annotation solves this.