add sop c16 for verifying external ocm sendgrid alerts
currently we will have alerts for the external ocm sendgrid service
that will not appear in the integreatly prometheus stack but in the
ocm prometheus stack.

this new sop ensures we check that alerts have been synced correctly
to the prometheus instance.

this verification only covers the staging environment, a future
commit will cover verifications of the alerts in the production
environment.

verification:
- follow c16