Reconcile Managed API addon subscription
The subscription controller decides whether to reconcile a subscription or not based
on the subscription namespace (if it matches the reconciler ns), and the name, if
it's either "integreatly", or "addon-rhmi" (The name of the subscription created by the
addon installation).
Refactor this logic to be based on a list of names to reconcile, and include
"addon-managed-api-service" to the list, in order to reconcile subscriptions created
by the Managed API addon.

Add unit tests for new logic