fix: Force rate limit configuration reload
The rate limit configuration config map would update the configuration file that's
mounted to the volume to the rate limit Pod. However, due to the mechanism to
mount ConfigMaps, the rate limit service wouldn't detect the changes and update
the configuration.

Update the rate limit reconciler to mount the ConfigMap as a file with a suffix
that's generated by hashing the rate limit configuration. This way, every time the
configuration changes, the Deployment will force to restart the Pod, loading the
updated configuration.