fix: Modify watches in namespace controller to enqueue correct requests
The namespace controller watches Namespaces and ConfigMaps, which will
enqueue different requests since the namespaces have their name in the
resource name, while the ConfigMap is namespaced. In order to ensure
uniform requests, add an EventHandler for the ConfigMap watch that
enqueues only its namespace, and add predicates for both types of
resources to watch exclusivelly what we're interested in: the operator
namespace and the add-on ConfigMaps.