fix: Fix make release/prepare
Fix `make release/prepare` to work with Kustomize when it's not
installed. The Makefile will automatically install (in a temporal
directory) Kustomize if it's not found on the $PATH. However, when
running `release/prepare`, it will invoke the `prepare-release.sh`
script, which assumes Kustomize is in the $PATH.
Modify the script to use an optional environment variable with the path
to Kustomize, and the Makefile to pass the temporal path when running
the script.