fix: Approve only latest InstallPlan
Currently, as part of the product operators installation, the
retrieval of the InstallPlans associated with the product is done by
listing all InstallPlans in the Subscription namespace. OLM however
might create multiple InstallPlans where only one is valid and is
referenced in the Subscription status as an `InstallPlanRef` field. The
current approach will loop through all the InstallPlans in the
namespace, approve them an wait for the status to be `Complete`, but if
an invalid InstallPlan comes first in the list, it will remain stuck
in progress, never reaching the valid one.

Change the `MarketPlaceInterface` and its `Manager` implementation to
retrieve only the InstallPlan that is referenced by the Subscription,
and update the unit tests that depend on this method.