feat!: Support "implicit" product operator installation
Extend product operator installation declaration file to support a new
type by specifying the `implicit` `installFrom` value.
When using this value for a product, the operator won't attempt to
reconcile a new CatalogSource for it, but instead will assume that the
CatalogSource that provided the RHOAM operator also provides this
product. Therefore, it will attempt to locate this CatalogSource and
create a Subscription for the product using that CatalogSource and the
specified `package` and `channel`