Mgdapi 2008 - Verify that user with uppercase letters can be created in 3scale (#1916)
* FIX: Upper case github user not being matched

It is not guaranteed that the keyclock user CR's will have lower case letters even with users created in side keyclock having lowercase. Every time the keyclock user CR username is checked it most be checked as lower case.

* Refactor: Move the lower casing of user names out of the function.

Some of the three scale client functions are lower casing the input of the username. This is misleading when debugging issues.

* Update: unit test to better match the use case.

As the keyclock users will always be lower case there, there can be no way to have two users with the same user name. The test has being updated to check that new users are found to be added.

* Feat: add new unit test case