fix(test): Introduce polling for SA token
The test has failed ocassionally when retrieving the token that's
created for a testing ServiceAccount. The reason of this is that the
token is attempted to be retrieved immediately after the SA is created,
which might result if the token not being created by the time that code
is executed. Fix by introducing polling which waits for the token secret
to be set in the service account `secrets` list.