Include suggested-namespace annotation in base CSV
The `operatorframework.io/suggested-namespace` is supported by OLM and
operator-sdk to specify the default/recommended namespace where the
operator might be installed. This provides a better user experience, as
well as advantages when testing using the scorecard format.

Include this annotation in the base CSVs of both RHOAM and RHMI. The
CSV generated will also include the annotation.